axf/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==1.1.0
billiard==3.3.0.23
cached-property==1.5.1
celery==3.1.25
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
Django==1.11.4
django-admin-view-permission==1.6
django-adminplus==0.5
django-braces==1.13.0
django-celery==3.2.1
django-debug-toolbar==2.0
django-extensions==1.8.1
django-oauth-toolkit==0.10.0
django-qiniu-storage==2.3.1
django-redis==4.8.0
django-redis-cache==2.1.0
django-url-filter==0.3.0
djangorestframework==3.4.6
djangorestframework-queryfields==1.0.0
djangorestframework-xml==1.3.0
drf-extensions==0.4.0
drf-tracking==1.3.1
enum34==1.1.6
gevent==1.1.2
greenlet==0.4.15
gunicorn==19.6.0
idna==2.8
jpush==3.3.0
kombu==3.0.37
lxml==4.3.1
ndg-httpsclient==0.5.1
oauthlib==1.0.3
Pillow==5.2.0
pyasn1==0.4.7
pycparser==2.19
pycryptodomex==3.7.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
python-alipay-sdk==1.10.1
python3-openid==3.1.0
pytz==2019.3
qiniu==7.1.0
redis==2.10.5
requests==2.22.0
requests-oauthlib==1.2.0
rest-social-auth==1.3.1
six==1.12.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.3.0
urllib3==1.25.6
uWSGI==2.0.18
